requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
dj-database-url==0.5.0
Django==4.0.2
django-filter==21.1
django-heroku==0.3.1
django-querycount==0.7.0
django-rest-framework==0.1.0
django-rest-knox==4.2.0
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
numpy==1.22.3
pandas==1.4.2
Pillow==9.0.1
psycopg2==2.9.3
pycparser==2.21
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
waitress==2.0.0
whitenoise==6.0.0
